tgz2build: hybrid linking on linux-glibc2.11
We have had a problem with Debian 10.
Our statically linked libc induced a segfault in the `time()` function.

It is not advised to link `libc` statically, so we create a hybrid
linkage:
 * ncurses and libm is linked statically
 * libc is linked dynamically

We already did this on solaris, but we do not have such a platform now,
and it seems it is necessary on linux platforms, too.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>